refactor: ui layout for design token consistency
