+++ view/app/self-host/components/github-connector/stepper-controls.tsx
@@ -55,4 +55,3 @@ export const StepperControls: React.FC<StepperControlsProps> = ({
     </Stepper.Controls>
   );
 };
-
