+++ view/app/self-host/components/github-connector/steps/install-app-step.tsx
@@ -29,4 +29,3 @@ export const InstallAppStep: React.FC<InstallAppStepProps> = ({
     </div>
   );
 };
-
